### 2020-07-28 #2532 @whedon
[PRE REVIEW]: SALSA: A Python Package for Constructing Synthetic Quasar Absorption Line Catalogs from Astrophysical Hydrodynamic Simulations
**Submitting author:** @biboyd (<a href="http://orcid.org/0000-0002-5419-9751">Brendan Boyd</a>)
**Repository:** <a href="https://github.com/biboyd/SALSA" target ="_blank">https://github.com/biboyd/SALSA</a>
**Version:** v0.1.0
