### 2020-07-30 cf0d31e
change absorber index to lightray index
