### 2020-08-05 #2532 @danielskatz
👋 @biboyd - please add an explicit section title Statement of Need to the paper - see https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain

Once you have done this, please enter ```@whedon generate pdf``` as a new command here to regenerate the pdf.
