### 2020-08-05 #2532 @danielskatz
Also, please suggest potential reviewers by mentioning them here in this thread (without tagging them with an @). [This list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list), or you may know of others in this field.  I will also check for reviewers through some other channels.
