### 2020-08-05 #2532 @danielskatz
@biboyd - I also notice that a number of your references seem to be missing journals - which in our style, makes them incomplete.  Please add the journals or conferences for all references.
