### 2020-08-10 #2532 @danielskatz
👋 @chummels & @brittonsmith - just checking with you again, and hoping that one of you will be willing to do this.

> 👋 @chummels & @brittonsmith - would one of you be willing to review this submission to JOSS? I am aware you are acknowledged in the paper, but I also think this is likely not a COI. Let me know what you think.
>
> If you're not familiar with JOSS's review style/guidelines, we are basically checklist driven and use GitHub, where reviewers, editors, and authors work openly and iteratively to ideally improve submissions to the point where they can be accepted. See https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
