### 2020-08-10 #2532 @danielskatz
👋 eteq & olebole - just checking with you again, and hoping that one of you will be willing to do this.

> 👋 eteq & olebole - would one of you be willing to review this JOSS submission?
