### 2020-08-10 #2532 @danielskatz
👋 @eteq & @olebole - would one of you be willing to review this JOSS submission?
