### 2020-08-10 #2532 @brittonsmith
@danielskatz, sorry for the delay in getting back to you. I work quite closely with the authors and attend regular meetings with them. I'm not sure if that is too close, but I am really drowning this summer. I just can't spare the time. My apologies. I can think of a couple people who would do an excellent job at this: neutrinoceros and cphyc. I hope they can help.
