### 2020-08-16 #2532 @danielskatz
👋 @cwchurchill-DrQSO, @gkacprzak, MTMurphy77 - would one of you be willing to review this submission to JOSS? You were suggested by a colleague at CfA.

If you're not familiar with JOSS's review style/guidelines, we are basically checklist driven and use GitHub, where reviewers, editors, and authors work openly and iteratively to ideally improve submissions to the point where they can be accepted. See https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
