### 2020-08-17 #2532 @gkacprzak
HI Daniel,

I am sorry, I just don't have the time. With full lockdown in Australia, no daycare, and two full time jobs are making it impossible to do this at this time.
Sorry about that.

Thank you,
Glenn

-----------------------------------------------------------

A/Prof. Glenn G. Kacprzak

Centre for Astrophysics & Supercomputing
Swinburne University of Technology
H29, PO Box 218,
Hawthorn, Victoria 3122, Australia
http://astronomy.swin.edu.au/~gkacprzak
Phone: +61 (0)3 9214 5439
Fax: +61 (0)3 9214 8797
-----------------------------------------------------------
