### 2020-08-18 #2581 @whedon
[REVIEW]: SALSA: A Python Package for Constructing Synthetic Quasar Absorption Line Catalogs from Astrophysical Hydrodynamic Simulations
**Submitting author:** @biboyd (<a href="http://orcid.org/0000-0002-5419-9751">Brendan Boyd</a>)
**Repository:** <a href="https://github.com/biboyd/SALSA" target ="_blank">https://github.com/biboyd/SALSA</a>
**Version:** v1.0.0
