### 2020-08-20 biboyd/SALSA#2 @olebole
[joss review] Citing yt in the paper?
Hi, I am reviewing the SALSA paper on Joss (openjournals/joss-reviews#2581). One question that arised for me is why the [yt software](https://yt-project.org/) is not cited? It seems to be used quite heavily.
