+++ paper.bib
@@ -36,6 +36,26 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Turk:2011,
+       author = {{Turk}, Matthew J. and {Smith}, Britton D. and {Oishi}, Jeffrey S. and
+         {Skory}, Stephen and {Skillman}, Samuel W. and {Abel}, Tom and
+         {Norman}, Michael L.},
+        title = "{yt: A Multi-code Analysis Toolkit for Astrophysical Simulation Data}",
+      journal = {\apjs},
+     keywords = {cosmology: theory, methods: data analysis, methods: numerical, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics},
+         year = 2011,
+        month = jan,
+       volume = {192},
+       number = {1},
+          eid = {9},
+        pages = {9},
+          doi = {10.1088/0067-0049/192/1/9},
+archivePrefix = {arXiv},
+       eprint = {1011.3514},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2011ApJS..192....9T},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
 
 @ARTICLE{Hummels:2017,
        author = {{Hummels}, Cameron B. and {Smith}, Britton D. and {Silvia}, Devin W.},
