### 2020-08-21 721df3d
specificy matplotlib3.3.1 since conflicts with yt
