### 2020-08-26 #2581 @danielskatz
👋 @zpace - it looks like you are close to finishing as well, other than the references checkbox

Is this correct?
