### 2020-08-26 #2581 @biboyd
@danielskatz I made a new release and archived on zenodo. Here is the information

Release: v1.0.0

zenodo archive: https://zenodo.org/record/4002068#.X0aJKxl7lhE

DOI: 10.5281/zenodo.4002067
