### 2020-08-26 b4d431a
minor grammar in JOSS paper
