+++ paper.md
@@ -78,20 +78,20 @@ data as well as help facilitate collaboration between simulators and observers.
 Two separate methods are made available to extract absorbers. The “Spectacle
 method” uses the Python package Spectacle[^2] to fit Voigt profiles to the
 synthetic spectra generated by Trident (see \autoref{fig:1}) [@spectacle]. This
-method provides traditional absorption line information (e.g. Doppler
-broadening, equivalent width, etc.) and thus creates synthetic absorber
+method provides traditional absorption line information (e.g., Doppler
+broadening, equivalent width) and thus creates synthetic absorber
 catalogs very similar to those made from observational studies.
 
 The second method, called the “SPICE (Simple Procedure for Iterative Cloud
-Extraction) method”, is a novel method that uses cell level data from the
+Extraction) method,” is a novel method that uses cell level data from the
 simulation to find the contiguous groups of cells which will meaningfully
 contribute absorption line features to the synthetic spectra. It does this
 through an iterative process that isolates the regions along the LightRay with
 the highest number density values and returns those regions with
 observationally detectable column densities as individual absorbers (see
 \autoref{fig:1} or for more details, the documentation[^3]). This method
-provides direct access to the information contained in the simulation (e.g.
-temperature, velocity, metallicity, etc.) and, in turn, more information than
+provides direct access to the information contained in the simulation (e.g.,
+temperature, velocity, metallicity) and, in turn, more information than
 can be provided by spectral absorption line analysis alone.
 
 ![These plots were generated using FOGGIE simulation data [@Peeples:2019]. The
