### 2020-08-26 #2581 @danielskatz
I found a few small fixes needed in the paper - https://github.com/biboyd/SALSA/pull/5

Otherwise, this looks ready to go to me

Please let me know when you have merged this, or what you disagree with
