### 2020-08-26 #2581 @biboyd
I merged in your small changes. I also reopened one of the reviewers issues https://github.com/biboyd/SALSA/issues/2 because it wasn't properly addressed. I just added a short snippet referencing the yt software in the paper and then edited the annotations accordingly.
