### 2020-08-26 #2581 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
