### 2020-08-27 #2581 @danielskatz
👋 @biboyd - sorry I forgot to ask you this before - is this submission associated with an AAS publication?
